refactor: extract store schema
